Moved file to specific folder, simplified initialisation, fixed namespace
